API_s.o.l.i.d_TS: migrate to a TypeORM DataSource, and turn build and test on

+++ API_s.o.l.i.d_TS/src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "../../../../shared/errors/AppError";
 import { deleteFile } from "../../../../utils/file";
 import { IUserRepository } from "../../repositories/IUserRepository";
 
@@ -12,6 +13,9 @@ class UpdateUserAvatarUseCase {
     constructor(@inject("UserRepository") private userRepository: IUserRepository) {}
     async execute({ userID, avatarFile }: IRequest): Promise<void> {
         const user = await this.userRepository.findById(userID);
+        if (!user) {
+            throw new AppError("User not found", 404);
+        }
         if (user.avatar) {
             await deleteFile(`./tmp/avatar/${user.avatar}`);
         }
